src/test/crimson: unify the two TestBlock implemenations
cache.cc is going to need a way to generate an appopriately
typed extent from a delta or on disk block based on the
type tag.  Thus, there can only be one TestBlock for now
unless it becomes at some point worth making it dynamic.
